refactor to move text elem code out of shop-layout
