+++ src/components/shop/confirm.js
@@ -12,8 +12,8 @@ import {
     createRect,
     getSafeArea,
     createPaneBackground,
-    addText,
 } from "./shop-layout.js";
+import { addText } from "../../core/layout/text-elem.js";
 import { createConfirmButtons } from "./menu-buttons.js";
 import { doTransaction } from "./transact.js";
 import { collections } from "../../core/collections.js";
