+++ src/core/layout/scrollable-list/button-overlays.js
@@ -5,7 +5,7 @@
  * @license Apache-2.0 Apache-2.0
  */
 import fp from "../../../../lib/lodash/fp/fp.js";
-import { addText } from "../../../components/shop/shop-layout.js";
+import { addText } from "../../../core/layout/text-elem.js";
 
 export const overlays1Wide = (gelButton, configs) => {
     configs.forEach(config => {
