+++ test/components/shop/shop-layout.test.js
@@ -236,38 +236,38 @@ describe("shop element scaling functions", () => {
         });
     });
 
-    describe("addText()", () => {
-        let elemConfig = { styles: { baz: "qux" } };
+    // describe("addText()", () => {
+    //     let elemConfig = { styles: { baz: "qux" } };
 
-        const mockScene = {
-            config: { styleDefaults: { foo: "bar" } },
-            add: { text: jest.fn().mockReturnValue("textElem") },
-        };
-        textUtils.updateStyleOnFontLoad = jest.fn();
+    //     const mockScene = {
+    //         config: { styleDefaults: { foo: "bar" } },
+    //         add: { text: jest.fn().mockReturnValue("textElem") },
+    //     };
+    //     textUtils.updateStyleOnFontLoad = jest.fn();
 
-        beforeEach(() => shopLayout.addText(mockScene, 0, 0, "someText", elemConfig));
+    //     beforeEach(() => shopLayout.addText(mockScene, 0, 0, "someText", elemConfig));
 
-        test("creates a text element", () => {
-            expect(mockScene.add.text).toHaveBeenCalled();
-        });
-        test("merges element styles with defaults", () => {
-            const expectedStyle = { foo: "bar", baz: "qux" };
-            expect(mockScene.add.text.mock.calls[0][3]).toStrictEqual(expectedStyle);
-        });
-        test("provides a fallback style", () => {
-            jest.clearAllMocks();
-            mockScene.config = {};
-            const fallbackStyle = {
-                fontFamily: "ReithSans",
-                fontSize: "24px",
-                resolution: 2,
-                align: "center",
-            };
-            shopLayout.addText(mockScene, 0, 0, "someText", undefined);
-            expect(mockScene.add.text.mock.calls[0][3]).toStrictEqual(fallbackStyle);
-        });
-        test("passes the text elem to updateStyleOnFontLoad", () => {
-            expect(textUtils.updateStyleOnFontLoad).toHaveBeenCalledWith("textElem");
-        });
-    });
+    //     test("creates a text element", () => {
+    //         expect(mockScene.add.text).toHaveBeenCalled();
+    //     });
+    //     test("merges element styles with defaults", () => {
+    //         const expectedStyle = { foo: "bar", baz: "qux" };
+    //         expect(mockScene.add.text.mock.calls[0][3]).toStrictEqual(expectedStyle);
+    //     });
+    //     test("provides a fallback style", () => {
+    //         jest.clearAllMocks();
+    //         mockScene.config = {};
+    //         const fallbackStyle = {
+    //             fontFamily: "ReithSans",
+    //             fontSize: "24px",
+    //             resolution: 2,
+    //             align: "center",
+    //         };
+    //         shopLayout.addText(mockScene, 0, 0, "someText", undefined);
+    //         expect(mockScene.add.text.mock.calls[0][3]).toStrictEqual(fallbackStyle);
+    //     });
+    //     test("passes the text elem to updateStyleOnFontLoad", () => {
+    //         expect(textUtils.updateStyleOnFontLoad).toHaveBeenCalledWith("textElem");
+    //     });
+    // });
 });
